controll remove user from chat only from owners

+++ src/components/User/User.jsx
@@ -1,28 +1,34 @@
 import { Button } from "@mui/joy";
-import React from "react";
+import React, { useContext } from "react";
 import { environment } from "../../environment/environment";
 import { useDispatch } from "react-redux/es/hooks/useDispatch";
 import { useSelector } from "react-redux/es/hooks/useSelector";
 import { setReduxStateDeleteUserFromChat } from "../../state/redux/actions/usersActions";
+import { useChatDetail } from "../../customHooks/useChatDetail";
+import { AuthContext } from "../../state/context/authContext";
 
 const User = ({ user }) => {
     const dispatch = useDispatch();
     const { chatActive } = useSelector((state) => state.chatActive);
+    const { userData } = useContext(AuthContext);
+    const chatDetail = useChatDetail(chatActive);
 
     const removeUser = async () => {
-        const bodyRequest = {
-            username: user.username,
-        };
-        const requestOptions = {
-            method: "DELETE",
-            headers: { "Content-Type": "application/json" },
-            body: JSON.stringify(bodyRequest),
-        };
-        await fetch(
-            `${environment.API_URL}/chat/remove/user/${chatActive.id}`,
-            requestOptions
-        );
-        dispatch(setReduxStateDeleteUserFromChat(user));
+        if (chatDetail.owner[0].username === userData.username) {
+            const bodyRequest = {
+                username: user.username,
+            };
+            const requestOptions = {
+                method: "DELETE",
+                headers: { "Content-Type": "application/json" },
+                body: JSON.stringify(bodyRequest),
+            };
+            await fetch(
+                `${environment.API_URL}/chat/remove/user/${chatDetail.id}`,
+                requestOptions
+            );
+            dispatch(setReduxStateDeleteUserFromChat(user));
+        }
     };
     return <Button onClick={removeUser}>{user.username}</Button>;
 };
